Move comparison code into its own module
